Llego a obtener un autovector y autovalor. Habria que hacer deflacion y encontrar los alpha-1 restantes. Igual estoy probando si esta funcionando

+++ tp2.data/src/Matrix.cpp
@@ -189,11 +189,11 @@ void Matrix::transpuesta() {
 	_transp = !_transp;
 }
 
-double& Matrix::operator()(const int& x, const int& y) {
+float& Matrix::operator()(const int& x, const int& y) {
 	return !_transp ? _mat[x][y] : _mat[y][x];
 }
 
-const double& Matrix::operator()(const int& x, const int& y)const {
+const float& Matrix::operator()(const int& x, const int& y)const {
 	return !_transp ? _mat[x][y] : _mat[y][x];
 }
