+++ tp2.data/src/Matrix.h
@@ -30,8 +30,8 @@ public:
 	vector<double> operator*(const vector<double>& b);
 	std::vector<double> diag_vec();
 
-	double& operator()(const int& n, const int& m);
-	const double& operator()(const int& n, const int& m) const;
+	float& operator()(const int& n, const int& m);
+	const float& operator()(const int& n, const int& m) const;
 
 	int getn() const;
 	int getm() const;
@@ -40,6 +40,6 @@ private:
 	int _n;
 	int _m;
 	bool _transp;
-	vector<vector<double>> _mat;
+	vector<vector<float>> _mat;
 };
